[WT-1238] move network tests to network.test.ts
